Safari-611.1.10/LayoutTests/fast/dom/document-all-tags-expected.txt

Test that HTMLAllCollection.tags() does not exist.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.all.tags is undefined.
PASS successfullyParsed is true

TEST COMPLETE
